chore: move current components to /classic
