+++ src/classic/components/Content/index.ts
