+++ src/classic/components/Footer/index.ts
