+++ src/classic/components/Header/Header.stories.tsx
@@ -1,6 +1,6 @@
 import { Story, Meta } from "@storybook/react"
 import Header, { HeaderProps } from "./Header"
-import Sitemap from "../../sitemap.json"
+import Sitemap from "../../../sitemap.json"
 
 export default {
   title: "Components/Header",
