+++ src/classic/components/Hero/Hero.tsx
