+++ src/classic/components/Hero/index.ts
