+++ src/classic/components/Image/Image.stories.tsx
