+++ src/classic/components/Image/Image.tsx
