+++ src/classic/components/Image/index.ts
