+++ src/classic/components/Infobar/index.ts
