+++ src/classic/components/Infopic/Infopic.tsx
