+++ src/classic/components/Infopic/index.ts
