+++ src/classic/components/Masthead/index.ts
