+++ src/classic/components/SidePane/SidePane.tsx
