+++ src/classic/components/SidePane/index.ts
