+++ src/layouts/Default/Default.tsx
@@ -1,4 +1,4 @@
-import { Header } from "../../components"
+import { Header } from "../../classic/components"
 import { IsomerBaseSchema, Sitemap, RenderEngine } from "../../engine/render"
 
 export interface DefaultLayoutProps {
